pyproject.toml
[project]
name = "bda-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.4",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-bigquery-storage>=2.36.2",
    "ipykernel>=7.2.0",
    "matplotlib>=3.10.8",
    "nbconvert>=7.17.0",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/gharchive", "src/ghrec"]
